refactor: 代码生成支持数据库类型为longtext
代码生成支持数据库类型为longtext

+++ src/main/java/com/youlai/boot/shared/codegen/model/bo/ColumnMetaData.java
@@ -25,7 +25,7 @@ public class ColumnMetaData {
         /**
          * 字段长度
          */
-        private Integer characterMaximumLength;
+        private Long characterMaximumLength;
 
         /**
          * 是否主键(1-是 0-否)
